requirements.txt
streamlit
supabase
pillow